+++ apps/decodex/src/agent/codex_accounts/record/line.rs
@@ -1,0 +1,100 @@
+use std::path::Path;
+
+use serde::{Deserialize, Serialize};
+
+use crate::{
+	agent::codex_accounts::record::{auth_json::AuthDotJson, model::AccountPoolRecord},
+	prelude::{Result, eyre},
+};
+
+#[derive(Clone, Deserialize, Serialize)]
+#[serde(untagged)]
+pub(super) enum AccountPoolLine {
+	Wrapped {
+		#[serde(skip_serializing_if = "Option::is_none")]
+		email: Option<String>,
+		#[serde(default, skip_serializing_if = "is_false")]
+		disabled: bool,
+		#[serde(skip_serializing_if = "Option::is_none")]
+		cooldown_until_unix_epoch: Option<i64>,
+		#[serde(skip_serializing_if = "Option::is_none")]
+		cooldown_until: Option<String>,
+		#[serde(skip_serializing_if = "Option::is_none")]
+		last_selected_at_unix_epoch: Option<i64>,
+		#[serde(skip_serializing_if = "Option::is_none")]
+		auth_failed_at_unix_epoch: Option<i64>,
+		#[serde(skip_serializing_if = "Option::is_none")]
+		auth_failure: Option<String>,
+		auth: AuthDotJson,
+	},
+	Flat(AccountPoolRecord),
+}
+impl AccountPoolLine {
+	pub(super) fn into_record(self) -> AccountPoolRecord {
+		match self {
+			Self::Flat(record) => record,
+			Self::Wrapped {
+				email,
+				disabled,
+				cooldown_until_unix_epoch,
+				cooldown_until,
+				last_selected_at_unix_epoch,
+				auth_failed_at_unix_epoch,
+				auth_failure,
+				auth,
+			} => AccountPoolRecord {
+				email: first_nonblank_string(email, auth.email),
+				disabled,
+				cooldown_until_unix_epoch,
+				cooldown_until,
+				last_selected_at_unix_epoch,
+				auth_failed_at_unix_epoch,
+				auth_failure,
+				auth_mode: auth.auth_mode,
+				openai_api_key: auth.openai_api_key,
+				tokens: auth.tokens,
+				last_refresh: auth.last_refresh,
+			},
+		}
+	}
+}
+
+pub(in crate::agent::codex_accounts) fn parse_account_records(
+	input: &str,
+	path: &Path,
+) -> Result<Vec<AccountPoolRecord>> {
+	let mut records = Vec::new();
+
+	for (line_index, line) in input.lines().enumerate() {
+		let line_number = line_index + 1;
+		let trimmed = line.trim();
+
+		if trimmed.is_empty() || trimmed.starts_with('#') {
+			continue;
+		}
+
+		let parsed = serde_json::from_str::<AccountPoolLine>(trimmed).map_err(|error| {
+			eyre::eyre!(
+				"Codex accounts `{}` line {line_number} is not a valid auth JSONL entry: {error}",
+				path.display()
+			)
+		})?;
+
+		records.push(parsed.into_record());
+	}
+
+	if records.is_empty() {
+		eyre::bail!("Codex accounts `{}` does not contain any account records.", path.display());
+	}
+
+	Ok(records)
+}
+
+pub(super) const fn is_false(value: &bool) -> bool {
+	!*value
+}
+
+fn first_nonblank_string(left: Option<String>, right: Option<String>) -> Option<String> {
+	left.filter(|value| !value.trim().is_empty())
+		.or_else(|| right.filter(|value| !value.trim().is_empty()))
+}
